% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeErrorMoment.R
\docType{methods}
\name{ComputeErrorMoment}
\alias{ComputeErrorMoment}
\alias{ComputeErrorMoment,contObsPredModelParam,AbsLossErrorMomentParam-method}
\title{\code{ComputeErrorMoments} computes the moments of measurement error for each statistical unit}
\usage{
ComputeErrorMoment(object, Param)


  \S4method{ComputeErrorMoment}{contObsPredModelParam,AbsLossErrorMomentParam}(object,
  Param)
}
\arguments{
\item{object}{Object of class \linkS4class{contObsPredModelParam} containing the statistical
units whose moments of measurement error for each variable is to be computed.}

\item{Param}{Object of virtual class \linkS4class{ErrorMomentParam} with the parameters determining
the method of computation of the moments of each statistical unit.}
}
\value{
Object of class \linkS4class{ErrorMoments} with the measurement error computed for each
variable and each statistical unit.
}
\description{
\code{ComputeErrorMoments} computes the moments of measurement error for each statistical unit
}
\examples{
\dontrun{

ImpParam <- new(Class = 'MeanImputationParam',
                VarNames = c('CifraNeg_13.___', 'Personal_07.__2.__'),
                DomainNames =  c('Tame_05._2.', 'ActivEcono_35._4._2.1.4._0'))

AbsLossPar <- new(Class = 'AbsLossErrorMomentParam',
                  VarNames =  c("CifraNeg_13.___", "Personal_07.__2.__"),
                  Homoskedastic = c(FALSE, FALSE),
                  Imputation = ImpParam)


ComputeErrorMoment(ObsPredPar, AbsLossPar)

}
}

