% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorMoments-class.R
\docType{class}
\name{ErrorMoments-class}
\alias{ErrorMoments-class}
\title{S4 class for the error moments in the optimization approach to selective editing}
\description{
Definition of the S4 class named \code{ErrorMoments} for the conditional moment  of
measurement errors in the optimization approach to selective editing.
}
\section{Slots}{

\describe{
\item{\code{VarNames}}{Character vector with the names of the variables whose error moments are contained
in this object.}

\item{\code{Domains}}{\linkS4class{data.table} with a row per domain with the values of the variables
determining the domains of the data set.}

\item{\code{Units}}{List of data.tables with as many components as domains (in the same order of the rows
of slot \code{Domains}) with every unit per domain.}

\item{\code{Moments}}{List of arrays with as many components as domains (in the same order of the rows
of slot \code{Domains}) with the set of error moments matrices per domain. Note that the order
of each error moment matrix is given by the order of the corresponding data.table in slot
\code{Units}.}
}}
\examples{
new(Class = 'ErrorMoments')

}

