% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbsLossErrorMomentParam-class.R
\docType{class}
\name{AbsLossErrorMomentParam-class}
\alias{AbsLossErrorMomentParam-class}
\title{S4 class for the parameters of the computation of moments of measurement error}
\description{
Definition of the S4 class named \code{AbsLossErrMomParam} for the parameters for
the computation of error moments using an absolute value loss function in the optimization
approach to selective editing.
}
\section{Slots}{

\describe{
\item{\code{VarNames}}{Character vector with the names of the variables whose probability errors are to
be computed.}

\item{\code{Homoskedastic}}{Logical}

\item{\code{UnitWeightNames}}{Character vector with the names of the variables representing the unit
weights in heteroskedastic observation-prediction models.}
}}

\examples{
AbsLossPar <- new(Class = 'AbsLossErrorMomentParam',
                  VarNames =  c("CifraNeg_13.___", "Personal_07._2.__"),
                  Homoskedastic = c(FALSE, FALSE))

}
